woopolemong_django/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
colorama==0.4.5
decorator==5.1.1
Django==3.2.13
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-extensions==3.2.0
django-imagekit==4.1.0
executing==1.0.0
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
prompt-toolkit==3.0.30
pure-eval==0.2.2
Pygments==2.13.0
pytz==2022.2.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.0
traitlets==5.3.0
wcwidth==0.2.5
